requirements.txt
feedparser
python-telegram-bot==20.3
python-dateutil
transformers
torch